fix the check for coverage
With the previous structure, if the coverage version was too old, the
raised exception for old version would never be seen. The enclosing
try-except block would catch it and the user would get the confusing
warning that coverage was not installed.

Instead, we get rid of the except and we check for the coverage
version outside of a try-except block. There's also no need to
except-else, since the original except block was raising again, and
thus the except-block would never be reached.